agent-workers/agent3/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
supabase==2.4.0
google-generativeai==0.5.0
python-dotenv==1.0.0
pydantic==2.7.0
resend==0.7.0
stripe==9.0.0

